# faster inv and div for Complex{Union{Float16, Float32}}