As discovered in https://discourse.julialang.org/t/efficient-calculation-of-many-complex-lorentzians-on-a-large-array/76150/21, Complex division for `Float16` and `Float32` are currently doing a lot of extra work. This still does the math in `Float64`, but ignores all the extra checks.