@oscardssmith 2022-02-10
Benchmarks look pretty good to me
```
#generates random numbers with widely varying values
myrand() = ComplexF32(reinterpret(Float32, rand(Int32)), reinterpret(Float32, rand(Int32)))
@btime myrand()
  3.528 ns (0 allocations: 0 bytes)

#old
@btime inv(myrand())  
10.503 ns (0 allocations: 0 bytes)

@btime myrand()/myrand()
  17.030 ns (0 allocations: 0 bytes)
  
#new
@btime inv(myrand())  
3.910 ns (0 allocations: 0 bytes)

@btime myrand()/myrand()
  6.893 ns (0 allocations: 0 bytes)
```