@LaurentPlagne 2022-02-10
Maybe I have missed something but isn't it a problem that CFloat32/16 semantics differs from CFloat64 ? (not the same cheks).