@giordano 2022-02-10
Can we please have correct functions rather than fast ones?
```julia
julia> function Base.inv(z::Complex{<:Union{Float16,Float32}})
           c, d = reim(widen(z))
           (isinf(c) | isinf(d)) && return complex(copysign(zero(z), c), flipsign(-zero(z), d))
           mag = inv(muladd(c, c, d^2))
           return oftype(z, Complex(c*mag, -d*mag))
       end

julia> inv(complex(Inf32))
ERROR: MethodError: no method matching copysign(::ComplexF32, ::Float64)
Closest candidates are:
  copysign(::Signed, ::Float64) at /usr/share/julia/base/int.jl:149
  copysign(::Signed, ::Real) at /usr/share/julia/base/int.jl:150
  copysign(::Rational, ::Real) at /usr/share/julia/base/rational.jl:256
  ...
Stacktrace:
 [1] inv(z::ComplexF32)
   @ Main ./REPL[1]:3
 [2] top-level scope
   @ REPL[3]:1
```