@giordano 2022-02-10
Relatedly, division of complex numbers has been plagued by several issues, some of them have been fixed over time, but for example #32992 still stands.  Might be a good occasion to fix it?