@oscardssmith 2022-02-10
Can you write me a test script for what the behavior should be? I think I want to merge this without fixing the Float64 edge cases (1.8 feature freeze is coming up), but I'd be happy to fix these in a follow-up PR.