@giordano 2022-02-10
```julia
@testset "Complex division by infinity - $T" for T in (Float16, Float32, Float64)
       @test isequal(complex(one(T)) / complex(T(Inf), T(Inf)), complex(zero(T)))
end
```

Note that with latest version of this PR `inv` of infinity is still broken:
```julia
julia> function Base.:/(z::Complex{T}, w::Complex{T}) where {T<:Union{Float16,Float32}}
           c, d = reim(widen(w))
           a, b = reim(widen(z))
           (isinf(c) | isinf(d)) && return complex(copysign(zero(T), c), flipsign(zero(T), d)) * z
           mag = inv(muladd(c, c, d^2))
           re_part = muladd(a, c, b*d)
           im_part = muladd(b, c, -a*d)
           return oftype(z, Complex(re_part*mag, im_part*mag))
       end

julia> function Base.inv(z::T) where T<:Complex{<:Union{Float16,Float32}}
           c, d = reim(widen(z))
           (isinf(c) | isinf(d)) && return complex(copysign(zero(T), c), flipsign(-zero(T), d))
           mag = inv(muladd(c, c, d^2))
           return oftype(z, Complex(c*mag, -d*mag))
       end

julia> inv(complex(Inf32))
ERROR: MethodError: no method matching copysign(::ComplexF32, ::Float64)
Closest candidates are:
  copysign(::Signed, ::Float64) at /usr/share/julia/base/int.jl:149
  copysign(::Signed, ::Real) at /usr/share/julia/base/int.jl:150
  copysign(::Rational, ::Real) at /usr/share/julia/base/rational.jl:256
  ...
Stacktrace:
 [1] inv(z::ComplexF32)
   @ Main ./REPL[5]:3
 [2] top-level scope
   @ REPL[6]:1
```
Same error as before