@oscardssmith 2022-02-11
@giordano how would you feel about changing the sign of the zeros for `a/b` to match those of `a*inv(b)`. Currently, `big(1.0 + 0.0im)/big(-Inf + Inf*im) != inv(big(-Inf + Inf*im))` which seems wrong.