@giordano 2022-02-11
What you suggest makes sense, but in v1.7.2 I get:
```julia
julia> big(1.0 + 0.0im)/big(-Inf + Inf*im) == inv(big(-Inf + Inf*im))
true
```
so that's already the case?  If in `master` you get something different that looks like a regression